Refactor form into controlled and uncontrolled variants
